Update FAQ.md
# FAQ

## What is Based Catchers?
Based Catcher Digital Collective (BCDC) is a gaming ecosystem powered by blockchain technology with NFTs as fun to collect as they are to play and interact with.

## Why are you making an NFT?
We are sick of what we've seen in the space.
/start rant
Disingenuous founders who talk of grand plans yet fail to deliver event the most mundane of achievements. The same overplayed, overhyped "frens." Coins? What coins. All the mindless cash grabs. "Sorry, it looks like our founder died." Royalty corruption by the web2 companies pretending to be "web3."
/end rant.
We think we can solve at least some of these problems by delivering a kick-ass ecosystem with tangible gaming value. Blockchain tech is amazing. Let's not waste any more of it on jpegs that should have stayed on Adobe.

## Who are the Founders?
Tom: Description; Adam: Description; Nikhil: Description (can click on the PFP and it will take them to our twitters)

## What are your plans for royalties?
Two things can be true. One, royalties attached to smart contracts are a great way to reward and compensate artists. Two, royalties have been abused by many profit-seeking web3 brands, and this expectation of easy money has led to a lack of innovation in the space. Because we do not identify as digital artists and will not rely on royalties to create more art and pay ourselves to eat, we feel a 1% royalty per transaction is appropriate. And we do not care if you trade on platforms that cut this to 0.5%.

## What are your plans for merch?
If we are to create merch, it will be to widen the scope of our brand to introduce our characters and game to a larger audience. Merch will likely be supplied at cost (or close to cost) to v1 Based Catcher holders.

## What are your plans to manage the floor price?
BCDC is a game on the chain. We don't care for nor do we manage floor prices. We'll let you weirdos worry about that.

## Where did the art come from?
Our art is a combination of ideas conceptualized by the founders, formulated by generative AI, sketched by several talented contractors, and brought to life by the one and only, Ian Bradley.